requirements.txt
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
datasets>=2.14.0
transformers>=4.30.0
tqdm>=4.65.0
matplotlib>=3.7.0
streamlit>=1.25.0
opencv-python-headless>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
gdown>=5.1.0
